Implemented decompress "Brotli" string when serialize.
