+++ HeadHunter.Database.MongoDb.Common/BsonSerializers/BrotliSerializer.cs
@@ -1,7 +1,6 @@
-﻿using MongoDB.Bson.Serialization;
+﻿using HeadHunter.Model.Common.Compression;
+using MongoDB.Bson.Serialization;
 using MongoDB.Bson.Serialization.Serializers;
-using System.IO.Compression;
-using System.Text;
 
 namespace HeadHunter.Database.MongoDb.Common.BsonSerializers
 {
@@ -9,37 +8,12 @@ namespace HeadHunter.Database.MongoDb.Common.BsonSerializers
     {
         public override void Serialize(BsonSerializationContext context, BsonSerializationArgs args, string value)
         {
-            var uncompressedText = value;
-            var uncompressedTextBytes = Encoding.UTF8.GetBytes(uncompressedText);
-
-            using (var outputStream = new MemoryStream())
-            {
-                using (var compressStream = new BrotliStream(outputStream, CompressionLevel.SmallestSize))
-                {
-                    compressStream.Write(uncompressedTextBytes, 0, uncompressedTextBytes.Length);
-                }
-
-                context.Writer.WriteString(Convert.ToBase64String(outputStream.ToArray()));
-            }
+            context.Writer.WriteString(BrotliCompressor.Compress(value));
         }
 
         public override string Deserialize(BsonDeserializationContext context, BsonDeserializationArgs args)
         {
-            var compressedText = context.Reader.ReadString();
-            var compressedTextBytes = Convert.FromBase64String(compressedText);
-
-            using (var inputStream = new MemoryStream(compressedTextBytes))
-            {
-                using (var outputStream = new MemoryStream())
-                {
-                    using (var decompressStream = new BrotliStream(inputStream, CompressionMode.Decompress))
-                    {
-                        decompressStream.CopyTo(outputStream);
-                    }
-
-                    return Convert.ToBase64String(outputStream.ToArray());
-                }
-            }
+            return BrotliCompressor.Decompress(context.Reader.ReadString());
         }
     }
 }
