+++ HeadHunter.Database.MongoDb.Common/HeadHunter.Database.MongoDb.Common.csproj
@@ -12,4 +12,8 @@
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\HeadHunter.Model.Common\HeadHunter.Model.Common.csproj" />
+  </ItemGroup>
+
 </Project>
