+++ HeadHunter.Database.MongoDb/Collections/Vacancy.cs
@@ -115,6 +115,7 @@ namespace HeadHunter.Database.MongoDb.Collections
 
         [BsonIgnoreIfNull]
         [BsonElement("description")]
+        [JsonConverter(typeof(BrotliConverter))]
         [BsonSerializer(typeof(BrotliSerializer))]
         public string? Description { get; set; }
 
@@ -128,6 +129,7 @@ namespace HeadHunter.Database.MongoDb.Collections
 
         [BsonIgnoreIfNull]
         [BsonElement("brandedDescription")]
+        [JsonConverter(typeof(BrotliConverter))]
         [BsonSerializer(typeof(BrotliSerializer))]
         public string? BrandedDescription { get; set; }
